dev: expose api
