+++ crates/tinymist-l10n/src/lib.rs
@@ -3,6 +3,8 @@
 use core::panic;
 use std::{
     borrow::Cow,
+    collections::HashSet,
+    path::Path,
     sync::{OnceLock, RwLock},
 };
 
@@ -62,6 +64,74 @@ pub fn load_translations(input: &str) -> anyhow::Result<TranslationMapSet> {
     });
 
     Ok(translations)
+}
+
+/// Updates disk translations with new key-value pairs.
+pub fn update_disk_translations(
+    mut key_values: Vec<(String, String)>,
+    output: &Path,
+) -> anyhow::Result<()> {
+    key_values.sort_by(|(key_x, _), (key_y, _)| key_x.cmp(key_y));
+
+    // Reads and parses existing translations
+    let mut translations = match std::fs::read_to_string(output) {
+        Ok(existing_translations) => deserialize(&existing_translations, true)?,
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => TranslationMapSet::default(),
+        Err(e) => Err(e)?,
+    };
+
+    // Removes unused translations
+    update_translations(key_values, &mut translations);
+
+    // Writes translations
+    let result = serialize_translations(translations);
+    std::fs::write(output, result)?;
+    Ok(())
+}
+
+/// Updates a map of translations with new key-value pairs.
+pub fn update_translations(
+    key_values: Vec<(String, String)>,
+    translations: &mut TranslationMapSet,
+) {
+    let used = key_values.iter().map(|e| &e.0).collect::<HashSet<_>>();
+    translations.retain(|k, _| used.contains(k));
+
+    // Updates translations
+    let en = "en".to_owned();
+    for (key, value) in key_values {
+        translations
+            .entry(key)
+            .or_default()
+            .insert(en.clone(), value);
+    }
+}
+
+/// Writes a map of translations to a TOML string.
+pub fn serialize_translations(translations: TranslationMapSet) -> String {
+    let mut result = String::new();
+
+    result.push_str("\n# The translations are partially generated by copilot\n");
+
+    let mut translations = translations.into_iter().collect::<Vec<_>>();
+    translations.sort_by(|a, b| a.0.cmp(&b.0));
+
+    for (key, mut data) in translations {
+        result.push_str(&format!("\n[{key}]\n"));
+
+        let en = data.remove("en").expect("en translation is missing");
+        result.push_str(&format!("en = {en}\n"));
+
+        // sort by lang
+        let mut data = data.into_iter().collect::<Vec<_>>();
+        data.sort_by(|a, b| a.0.cmp(&b.0));
+
+        for (lang, value) in data {
+            result.push_str(&format!("{lang} = {value}\n"));
+        }
+    }
+
+    result
 }
 
 /// Tries to translate a string to the current language.
