+++ crates/tinymist-l10n/src/main.rs
@@ -1,13 +1,13 @@
 //! Fully parallelized l10n tool for Rust and TypeScript.
 
-use std::collections::HashSet;
+use std::path::Path;
 
 use clap::Parser;
 use rayon::{
     iter::{ParallelBridge, ParallelIterator},
     str::ParallelString,
 };
-use tinymist_l10n::{deserialize, TranslationMapSet};
+use tinymist_l10n::update_disk_translations;
 
 /// The CLI arguments of the tool.
 #[derive(Debug, Clone, PartialEq, Eq, Parser)]
@@ -32,7 +32,7 @@ fn main() -> anyhow::Result<()> {
     let args = Args::parse();
 
     let is_rs = args.kind == "rs";
-    let mut file_calls = walkdir::WalkDir::new(&args.dir)
+    let file_calls = walkdir::WalkDir::new(&args.dir)
         .into_iter()
         .filter_map(|e| e.ok())
         .filter(|e| e.file_type().is_file())
@@ -45,59 +45,9 @@ fn main() -> anyhow::Result<()> {
         .flat_map(|e| check_calls(e, is_rs))
         .collect::<Vec<_>>();
 
-    file_calls.sort_by(|(key_x, _), (key_y, _)| key_x.cmp(key_y));
-
-    // Reads and parses existing translations
-    let mut translations = match std::fs::read_to_string(&args.output) {
-        Ok(existing_translations) => deserialize(&existing_translations, true)?,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => TranslationMapSet::default(),
-        Err(e) => Err(e)?,
-    };
-
-    // Removes unused translations
-    let used = file_calls.iter().map(|e| &e.0).collect::<HashSet<_>>();
-    translations.retain(|k, _| used.contains(k));
-
-    // Updates translations
-    let en = "en".to_owned();
-    for (key, value) in file_calls {
-        translations
-            .entry(key)
-            .or_default()
-            .insert(en.clone(), value);
-    }
-
-    // Writes translations
-    let result = write_toml(translations);
-    std::fs::write(&args.output, result)?;
+    update_disk_translations(file_calls, Path::new(&args.output))?;
 
     Ok(())
-}
-
-fn write_toml(translations: TranslationMapSet) -> String {
-    let mut result = String::new();
-
-    result.push_str("\n# The translation are partially generated by copilot\n");
-
-    let mut translations = translations.into_iter().collect::<Vec<_>>();
-    translations.sort_by(|a, b| a.0.cmp(&b.0));
-
-    for (key, mut data) in translations {
-        result.push_str(&format!("\n[{key}]\n"));
-
-        let en = data.remove("en").expect("en translation is missing");
-        result.push_str(&format!("en = {en}\n"));
-
-        // sort by lang
-        let mut data = data.into_iter().collect::<Vec<_>>();
-        data.sort_by(|a, b| a.0.cmp(&b.0));
-
-        for (lang, value) in data {
-            result.push_str(&format!("{lang} = {value}\n"));
-        }
-    }
-
-    result
 }
 
 const L10N_FN_TS: &str = "l10nMsg";
